Festival Manager Authorization
for Edit, Create and Delete

+++ FestMVC/Controllers/EventsController.cs
@@ -20,13 +20,10 @@ namespace FestMVC.Controllers
         // GET: Events
         public ActionResult Index()
         {
-            var events = db.Events.Include(f => f.Festival).Include(f => f.Instructor).Include(f => f.Room).ToList<IbaseModel>();
-
-            if (User.IsInRole("FestivalManager"))
-            {
-                events = Utilities.FilterFestivalManager(events, User.Identity.GetUserId());
-            }
-
+            
+            var events = Utilities.FilterFestivalManager(
+                db.Events.Include(f => f.Festival).Include(f => f.Instructor).Include(f => f.Room).ToList<IbaseModel>(), User);
+           
             return View(events.OfType<Event>());
         }
 
@@ -71,8 +68,8 @@ namespace FestMVC.Controllers
         }
 
         // GET: Events/Create
-        [OverrideAuthorization]
-        [Authorize(Roles = "Admin, FestivalManager, User")]
+        //[OverrideAuthorization]
+        //[Authorize(Roles = "Admin, FestivalManager, User")]
         public ActionResult Create()
         {
             //ViewBag.FestivalId = new SelectList(db.Festivals, "Id", "Name");
@@ -86,8 +83,8 @@ namespace FestMVC.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        [OverrideAuthorization]
-        [Authorize(Roles = "Admin, FestivalManager, User")]
+        //[OverrideAuthorization]
+        //[Authorize(Roles = "Admin, FestivalManager, User")]
         public ActionResult Create([Bind(Include = "Id,Name,Description,StartDate,StartTime,EndTime,FestivalId,InstructorId,RoomId")] EventViewModel @eventViewModel)
         {
             if (ModelState.IsValid)
@@ -95,7 +92,13 @@ namespace FestMVC.Controllers
                 Event @event = new Event(@eventViewModel.Id, @eventViewModel.Name, @eventViewModel.Description, @eventViewModel.FestivalId,
                     @eventViewModel.InstructorId, @eventViewModel.RoomId,
                     @eventViewModel.StartDate, @eventViewModel.StartTime, @eventViewModel.EndTime);
-                //Event @event = new Event();
+
+                Festival festival = db.Festivals.Find(@event.FestivalId);
+
+                if (Utilities.IsForbidden(festival, User))
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.Forbidden);
+                }
 
                 db.Events.Add(@event);
                 db.SaveChanges();
@@ -115,11 +118,21 @@ namespace FestMVC.Controllers
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
+
             Event @event = db.Events.Find(id);
+
             if (@event == null)
             {
                 return HttpNotFound();
             }
+
+            Festival festival = db.Festivals.Find(@event.FestivalId);
+
+            if (Utilities.IsForbidden(festival, User))
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.Forbidden);
+            }
+
             EventViewModel @eventViewModel = new EventViewModel(@event.Id, @event.Name, @event.Description, @event.FestivalId,
                     @event.InstructorId, @event.RoomId, @event.StartDate, @event.EndDate, @event.Festival, @event.Room, @event.Instructor);
             //ViewBag.FestivalId = new SelectList(db.Festivals, "Id", "Name", @event.FestivalId);
@@ -141,6 +154,13 @@ namespace FestMVC.Controllers
                 Event @event = new Event(@eventViewModel.Id, @eventViewModel.Name, @eventViewModel.Description, @eventViewModel.FestivalId,
                     @eventViewModel.InstructorId, @eventViewModel.RoomId,
                     @eventViewModel.StartDate, @eventViewModel.StartTime, @eventViewModel.EndTime);
+
+                Festival festival = db.Festivals.Find(@event.FestivalId);
+
+                if (Utilities.IsForbidden(festival, User))
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.Forbidden);
+                }
 
                 db.Entry(@event).State = EntityState.Modified;
                 db.SaveChanges();
@@ -165,6 +185,12 @@ namespace FestMVC.Controllers
             {
                 return HttpNotFound();
             }
+
+            if (Utilities.IsForbidden(@event, User))
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.Forbidden);
+            }
+
             EventViewModel @eventViewModel = new EventViewModel(@event.Id, @event.Name, @event.Description, @event.FestivalId,
                     @event.InstructorId, @event.RoomId, @event.StartDate, @event.EndDate, @event.Festival, @event.Room, @event.Instructor);
             return View(@eventViewModel);
@@ -176,6 +202,12 @@ namespace FestMVC.Controllers
         public ActionResult DeleteConfirmed(long id)
         {
             Event @event = db.Events.Find(id);
+
+            if (Utilities.IsForbidden(@event, User))
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.Forbidden);
+            }
+
             db.Events.Remove(@event);
             db.SaveChanges();
             return RedirectToAction("Index");
@@ -191,13 +223,16 @@ namespace FestMVC.Controllers
         }
         private void PopulateDropDownList(object selectedUser = null, object selectedFestival = null, object selectedRoom = null)
         {
+            
+            var festivals = Utilities.FilterFestivalManager(db.Festivals.Include(f => f.FestivalManager).ToList<IbaseModel>(), User);
+           
             var usersQuery = from d in db.Instructors
                              orderby d.User.Name
                              select (new { d.Id, d.User.Name });
-            ViewBag.InstructorId = new SelectList(usersQuery, "Id", "Name", selectedUser);
-            ViewBag.FestivalId = new SelectList(db.Festivals, "Id", "Name", selectedFestival);
-            ViewBag.RoomId = new SelectList(db.Rooms, "Id", "Name", selectedRoom);
 
+            ViewBag.InstructorId = new SelectList(usersQuery, "Id", "Name", selectedUser);
+            ViewBag.FestivalId = new SelectList(festivals.OfType<Festival>(), "Id", "Name", selectedFestival);
+            ViewBag.RoomId = new SelectList(db.Rooms, "Id", "Name", selectedRoom);
 
         }
     }
